+++ TOMB4/game/health.h
@@ -10,7 +10,7 @@ void DrawPickups();
 void AddDisplayPickup(short object_number);
 
 extern DISPLAYPU pickups[8];
-extern short PickupX;
+extern long PickupX;
 extern short CurrentPickup;
 
 extern long health_bar_timer;